KVM: selftests: Add dirty ring buffer test
Add the initial dirty ring buffer test.

The current test is a simplest way of implementing a userspace dirty
ring collection mechanism, by only reaping the dirty ring when the
ring is full.

So it's still as simple as:

            vcpu                                 main thread

  1. vcpu dirties pages
  2. vcpu gets dirty ring full
     (userspace exit)
                                           3. main thread waits until full
                                              (so hardware buffers flushed)
                                           4. main thread collects
                                           5. main thread continues vcpu
  6. vcpu continues, goes back to 1

We can't directly collects dirty bits because otherwise we can't
guarantee the hardware dirty bits were flushed when we collect.  A
follow up patch will make it to kick vcpu and finally support async.

Signed-off-by: Peter Xu <peterx@redhat.com>