KVM: selftests: Let dirty_log_test async for dirty ring test
Previously the dirty ring test was working in synchronous way, because
only with a vmexit (with that it was the ring full event) we'll know
the hardware dirty bits will be flushed to the dirty ring.

With this patch we first introduced the vcpu kick mechanism by using
SIGUSR1, meanwhile we can have a guarantee of vmexit and also the
flushing of hardware dirty bits.  With all these, we can keep the vcpu
dirty work asynchronous of the whole collection procedure now.

Further increase the dirty ring size to current maximum to make sure
we torture more on the no-ring-full case, which can be mostly async
like this.

Signed-off-by: Peter Xu <peterx@redhat.com>