kthread: work could not be queued when worker being destroyed
The "queuing_blocked" func should print warning message and returns true
when the worker being destroyed.

Before the work is put into the queue of the worker thread, the state of
the worker thread needs to be detected,because the worker thread may be in
the destruction state at this time.

Link: http://lkml.kernel.org/r/20200705013018.7375-1-qiang.zhang@windriver.com
Link: http://lkml.kernel.org/r/20200702070156.5862-1-qiang.zhang@windriver.com
Signed-off-by: Zhang Qiang <qiang.zhang@windriver.com>
Suggested-by: Petr Mladek <pmladek@suse.com>
Reviewed-by: Petr Mladek <pmladek@suse.com>
Cc: Tejun Heo <tj@kernel.org>
Cc: Ben Dooks (Codethink) <ben.dooks@codethink.co.uk>
Cc: J. Bruce Fields <bfields@redhat.com>
Cc: Liang Chen <cl@rock-chips.com>
Cc: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>