hugetlb/userfaultfd: Unshare all pmds for hugetlbfs when register
Huge pmd sharing for hugetlbfs is racy with userfaultfd because userfaultfd
information is always based on pgtable entries, so they cannot be shared.

Walk the hugetlb range and unshare all such mappings if there is, right before
UFFDIO_REGISTER will succeed and return to userspace.

This will pair with want_pmd_share() in hugetlb code so that huge pmd sharing
is completely disabled for userfaultfd registered range.

Signed-off-by: Peter Xu <peterx@redhat.com>