btrfs: zoned: defer loading zone info after opening trees
This is preparation patch to implement zone emulation on a regular device.

To emulate zoned mode on a regular (non-zoned) device, we need to decide
an emulating zone size. Instead of making it compile-time static value,
we'll make it configurable at mkfs time. Since we have one zone == one
device extent restriction, we can determine the emulated zone size from
the size of a device extent. We can extend btrfs_get_dev_zone_info() to
show a regular device filled with conventional zones once the zone size
is decided.

The current call site of btrfs_get_dev_zone_info() during the mount
process is earlier than reading the trees, so we can't slice a regular
device to conventional zones. This patch defers the loading of zone info
to open_ctree() to load the emulated zone size from a device extent.

Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>
Signed-off-by: David Sterba <dsterba@suse.com>