btrfs: zoned: use regular super block location in emulated zoned mode
The zoned mode puts a superblock at the beginning of SB logging zones if
the zone is conventional. This difference causes a chicken-and-egg
problem for emulated zoned mode. Since the device is a regular
(non-zoned) device, we cannot know if the filesystem is regular or
emulated zoned while we read the superblock. But, to load proper
superblock, we need to see if it is emulated zoned or not.

We place the super blocks at the same location as the regular btrfs on
emulated zoned mode to solve the problem. It is possible because it's
ensured that all the SB locations are at a conventional zone on emulated
zoned mode.

Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>
Signed-off-by: David Sterba <dsterba@suse.com>