btrfs: zoned: do not load fs_info::zoned from incompat flag
Don't set the zoned flag in fs_info when encountering the incompat ZONED
on mount. The zoned flag in fs_info is in a union together with the
zone_size, so setting it too early will result in setting an incorrect
zone_size as well.

Once the correct zone_size is read from the device, we can rely on the
zoned flag in fs_info as well to determine if the filesystem is running in
zoned mode.

Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>