btrfs: zoned: implement zoned chunk allocator
Implement a zoned chunk/dev_extent allocator. The device extents are
aligned to the zone boundaries, so that a zone reset affects only the
device extent and does not change the state of blocks in the neighbor
device extents.

Also, it checks that a region allocation is not overlapping any of the
super block zones, and ensures the region is empty.

Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>
Signed-off-by: David Sterba <dsterba@suse.com>