btrfs: zoned: implement sequential extent allocation
Implements a sequential extent allocator for the zoned mode.  This
allocator just needs to check if there is enough space in the block
group. Therefore the allocator never manages bitmaps or clusters. Also
add assertions to the corresponding functions.

Actually, with zone append writing, it is unnecessary to track the
allocation offset. It only needs to check space availability. But, by
tracking the offset and returning the offset as an allocated region, we
can skip modification of ordered extents and checksum information when
there is no IO reordering.

Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>
Signed-off-by: David Sterba <dsterba@suse.com>