btrfs: factor out helper adding a page to bio
This commit extract page adding to bio part from submit_extent_page().
The page is added only when bio_flags are the same, contiguous and the
added page fits in the same stripe as pages in the bio.

Condition checks are reordered to allow early return to avoid possibly
heavy btrfs_bio_fits_in_stripe() calling.

Reviewed-by: Josef Bacik <josef@toxicpanda.com>
Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>
Signed-off-by: David Sterba <dsterba@suse.com>