userfaultfd/selftests: Remove the time() check on delayed uffd
There seems to have no guarantee that time() will return the same for the two
calls even if there's no delay, e.g. when a fault is accidentally crossing the
changing of a second.  Meanwhile, this message is also not helping that much
since delay could happen with a lot of reasons, e.g., schedule latency of
resolving thread.  It may not mean an issue with uffd.

Neither do I saw this error triggered either in the past runs.  Even if it
triggers, it'll be drown in all the rest of test logs.  Remove it.

Reviewed-by: Axel Rasmussen <axelrasmussen@google.com>
Signed-off-by: Peter Xu <peterx@redhat.com>